fix: implement game files into an existing file structure
