--- gameboard/game.css
@@ -1,245 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Lato:wght@400;700&display=swap");
-* {
-  margin: 0;
-  padding: 0;
-  -webkit-box-sizing: border-box;
-          box-sizing: border-box;
-  color: #fff;
-  font-family: "Lato", sans-serif;
-}
-
-.wrapper {
-  height: 100vh;
-  width: 100vw;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  background-image: url("./bg.png");
-}
-
-.wrapper__main {
-  background-color: rgba(0, 0, 0, 0.45);
-  height: 100%;
-  width: 90%;
-  margin: 0 auto;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  position: relative;
-}
-
-.timer {
-  position: absolute;
-  top: 5%;
-}
-
-.timer__counter {
-  font-size: 48px;
-  font-weight: bold;
-}
-
-.gameBoard__wrapper {
-  height: 100%;
-  width: 100%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  position: relative;
-}
-
-.gameBoard__player {
-  position: absolute;
-  top: 10%;
-}
-
-.gameBoard__player-name {
-  font-size: 40px;
-}
-
-.gameBoard__board {
-  width: 400px;
-  height: 400px;
-  background-color: #fff;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-      flex-wrap: wrap;
-  position: relative;
-}
-
-.gameBoard__board-fieldsWrapper {
-  height: 100%;
-  width: 100%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-      flex-wrap: wrap;
-}
-
-.gameBoard__board-field {
-  height: 10%;
-  width: 10%;
-  border: 1px solid black;
-}
-
-.gameBoard__board-field:nth-child(-n + 10) {
-  border-top: none;
-}
-
-.gameBoard__board-field:nth-child(10n) {
-  border-right: none;
-}
-
-.gameBoard__board-field:nth-child(10n + 1) {
-  border-left: none;
-}
-
-.gameBoard__board-field:nth-child(n + 91) {
-  border-bottom: none;
-}
-
-.gameBoard__RandomizeButton {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  -webkit-transform: translate(-50%, -50%);
-          transform: translate(-50%, -50%);
-  height: 25%;
-  width: 85%;
-  background-color: #243c60;
-  border: none;
-  border-radius: 15px;
-  font-size: 48px;
-  font-weight: 700;
-}
-
-.gameBoard__RandomizeButton:hover {
-  cursor: pointer;
-  background-color: #051e42;
-}
-
-.gameBoard__charactersWrapper {
-  position: absolute;
-  width: 40px;
-  height: 100%;
-  left: -40px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-}
-
-.gameBoard__charactersWrapper-char {
-  height: 40px;
-  width: 40px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-}
-
-.gameBoard__charactersWrapper-char h1 {
-  font-size: 24px;
-}
-
-.gameBoard__numbersWrapper {
-  position: absolute;
-  width: 100%;
-  height: 40px;
-  top: -40px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-}
-
-.gameBoard__numbersWrapper-numb {
-  height: 50px;
-  width: 50px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-}
-
-.gameBoard__numbersWrapper-numb h1 {
-  font-size: 24px;
-}
-
-.gameBoard__shipsWrapper {
-  position: absolute;
-  bottom: 5%;
-  width: 400px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-      flex-wrap: wrap;
-}
-
-.gameBoard__shipsWrapper-ships {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-      flex-wrap: wrap;
-  -webkit-box-pack: start;
-      -ms-flex-pack: start;
-          justify-content: start;
-  -webkit-box-align: end;
-      -ms-flex-align: end;
-          align-items: end;
-}
-
-.gameBoard__shipsWrapper-ship {
-  background-color: #fff;
-  height: 25px;
-  margin: 5px 10px 5px 0;
-}
-
-.gameBoard__shipsWrapper__heading {
-  font-weight: 700;
-}
-
-.homaPageButton {
-  border: none;
-  background-color: transparent;
-  color: #fff;
-  font-size: 24px;
-  font-weight: 700;
-  position: absolute;
-  right: 5%;
-  bottom: 3%;
-}
-
-.homaPageButton:hover {
-  cursor: pointer;
-}
-/*# sourceMappingURL=game.css.map */
