--- gameboard/setupships.css
@@ -1,275 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Lato:wght@400;700&display=swap");
-* {
-  margin: 0;
-  padding: 0;
-  -webkit-box-sizing: border-box;
-          box-sizing: border-box;
-  color: #fff;
-  font-family: "Lato", sans-serif;
-}
-
-.wrapper {
-  height: 100vh;
-  width: 100vw;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  background-image: url("bg.png");
-}
-
-.wrapper__main {
-  background-color: rgba(0, 0, 0, 0.45);
-  height: 100%;
-  width: 90%;
-  margin: 0 auto;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  position: relative;
-}
-
-.wrapper__header {
-  position: absolute;
-  top: 10%;
-  left: 50%;
-  -webkit-transform: translate(-50%, -50%);
-          transform: translate(-50%, -50%);
-}
-
-.gameBoard__wrapper {
-  height: 100%;
-  width: 100%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  position: relative;
-}
-
-.gameBoard__board {
-  width: 400px;
-  height: 400px;
-  background-color: #fff;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-      flex-wrap: wrap;
-  position: relative;
-  margin: 0 50px 0 0;
-}
-
-.gameBoard__board-fieldsWrapper {
-  height: 100%;
-  width: 100%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-      flex-wrap: wrap;
-}
-
-.gameBoard__board-field {
-  height: 10%;
-  width: 10%;
-  border: 1px solid black;
-}
-
-.gameBoard__board-field:nth-child(-n + 10) {
-  border-top: none;
-}
-
-.gameBoard__board-field:nth-child(10n) {
-  border-right: none;
-}
-
-.gameBoard__board-field:nth-child(10n + 1) {
-  border-left: none;
-}
-
-.gameBoard__board-field:nth-child(n + 91) {
-  border-bottom: none;
-}
-
-.gameBoard__RandomizeButton {
-  height: 20%;
-  width: 100%;
-  background-color: #243c60;
-  border: none;
-  border-radius: 15px;
-  font-size: 48px;
-  font-weight: 700;
-  -ms-flex-item-align: end;
-      align-self: flex-end;
-  -webkit-transform: translateY(40px);
-          transform: translateY(40px);
-}
-
-.gameBoard__RandomizeButton:hover {
-  cursor: pointer;
-  background-color: #051e42;
-}
-
-.gameBoard__charactersWrapper {
-  position: absolute;
-  width: 40px;
-  height: 100%;
-  left: -40px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-}
-
-.gameBoard__charactersWrapper-char {
-  height: 40px;
-  width: 40px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-}
-
-.gameBoard__charactersWrapper-char h1 {
-  font-size: 24px;
-}
-
-.gameBoard__numbersWrapper {
-  position: absolute;
-  width: 100%;
-  height: 40px;
-  top: -40px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-}
-
-.gameBoard__numbersWrapper-numb {
-  height: 50px;
-  width: 50px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-}
-
-.gameBoard__numbersWrapper-numb h1 {
-  font-size: 24px;
-}
-
-.gameBoard__shipsWrapper {
-  width: 400px;
-  height: 400px;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-      flex-wrap: wrap;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  -webkit-transform: translateY(-40px);
-          transform: translateY(-40px);
-}
-
-.gameBoard__shipsWrapper-setupShips {
-  width: 100%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-      flex-wrap: wrap;
-  -webkit-box-align: start;
-      -ms-flex-align: start;
-          align-items: flex-start;
-}
-
-.gameBoard__shipsWrapper-ship {
-  background-color: #fff;
-  height: 40px;
-  margin: 5px 10px 5px 0;
-}
-
-.gameBoard__shipsWrapper__heading {
-  font-weight: 700;
-  width: 100%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-}
-
-.gameBoard__rotateButton {
-  height: 15%;
-  width: 30%;
-  background-color: #243c60;
-  border: none;
-  border-radius: 15px;
-  font-size: 24px;
-  font-weight: 700;
-  -ms-flex-item-align: end;
-      align-self: flex-end;
-  -webkit-transform: translateY(40px);
-          transform: translateY(40px);
-}
-
-.gameBoard__rotateButton:hover {
-  cursor: pointer;
-  background-color: #051e42;
-}
-
-.homaPageButton {
-  border: none;
-  background-color: transparent;
-  color: #fff;
-  font-size: 24px;
-  font-weight: 700;
-  position: absolute;
-  right: 5%;
-  bottom: 3%;
-}
-
-.homaPageButton:hover {
-  cursor: pointer;
-}
-
-.startGameButton {
-  height: 10%;
-  width: 20%;
-  background-color: #614029;
-  border: none;
-  border-radius: 15px;
-  font-size: 24px;
-  font-weight: 700;
-  position: absolute;
-  bottom: 3%;
-}
-
-.startGameButton:hover {
-  cursor: pointer;
-  background-color: #ad7751;
-}
-/*# sourceMappingURL=setupships.css.map */
