--- gameboard/setupships.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="setupships.css">
-    <title>Document</title>
-</head>
-<body>
-    <div class="wrapper">
-        <div class="wrapper__main">
-<h1 class="wrapper__header">Set your ships</h1>
-            <div class="gameBoard__wrapper">
-                <div class="gameBoard__board">
-                    <div class="gameBoard__board-fieldsWrapper"></div>
-                    <div class="gameBoard__charactersWrapper"></div>
-                    <div class="gameBoard__numbersWrapper"></div>
-                </div>
-                <div class="gameBoard__shipsWrapper">
-                    <h1 class="gameBoard__heading">Ships left:</h1>
-                    <div class="gameBoard__shipsWrapper-setupShips"></div>
-                    <button class="gameBoard__rotateButton">Rotate</button>
-                    <button class="gameBoard__RandomizeButton">Randomize</button>
-                </div>
-            </div>
-            
-            <button class="startGameButton">Start Game</button>
-            <button class="homaPageButton">Return</button>
-        </div>
-    </div>
-
-    <script src="setupships.js"></script>
-</body>
-</html>
