--- gameboard/setupships.scss
@@ -1,221 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Lato:wght@400;700&display=swap");
-
-$base-font-color: #fff;
-$btn-color: #243c60;
-$bgc-color: rgba(0, 0, 0, 0.45);
-$btn-color-hover: #051e42;
-$base-font-family: "Lato", sans-serif;
-$bg-url: url("bg.png");
-
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  color: $base-font-color;
-  font-family: $base-font-family;
-}
-
-.wrapper {
-  height: 100vh;
-  width: 100vw;
-  display: flex;
-  background-image: $bg-url;
-
-  &__main {
-    background-color: $bgc-color;
-    height: 100%;
-    width: 90%;
-    margin: 0 auto;
-    display: flex;
-    justify-content: center;
-    position: relative;
-  }
-
-  &__header {
-    position: absolute;
-    top: 10%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-}
-.gameBoard {
-  &__wrapper {
-    height: 100%;
-    width: 100%;
-    display: flex;
-    // flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-  }
-
-  &__board {
-    width: 400px;
-    height: 400px;
-    background-color: #fff;
-    display: flex;
-    flex-wrap: wrap;
-    position: relative;
-    margin: 0 50px 0 0;
-
-    &-fieldsWrapper {
-      height: 100%;
-      width: 100%;
-      display: flex;
-      flex-wrap: wrap;
-    }
-
-    &-field {
-      height: 10%;
-      width: 10%;
-      border: 1px solid black;
-      &:nth-child(-n + 10) {
-        border-top: none;
-      }
-
-      &:nth-child(10n) {
-        border-right: none;
-      }
-
-      &:nth-child(10n + 1) {
-        border-left: none;
-      }
-
-      &:nth-child(n + 91) {
-        border-bottom: none;
-      }
-    }
-  }
-  &__RandomizeButton {
-    height: 20%;
-    width: 100%;
-    background-color: $btn-color;
-    border: none;
-    border-radius: 15px;
-    font-size: 48px;
-    font-weight: 700;
-    align-self: flex-end;
-    transform: translateY(40px);
-
-    &:hover {
-      cursor: pointer;
-      background-color: $btn-color-hover;
-    }
-  }
-  &__charactersWrapper {
-    position: absolute;
-    width: 40px;
-    height: 100%;
-    left: -40px;
-    display: flex;
-    flex-direction: column;
-
-    &-char {
-      height: 40px;
-      width: 40px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      h1 {
-        font-size: 24px;
-      }
-    }
-  }
-
-  &__numbersWrapper {
-    position: absolute;
-    width: 100%;
-    height: 40px;
-    top: -40px;
-    display: flex;
-
-    &-numb {
-      height: 50px;
-      width: 50px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      h1 {
-        font-size: 24px;
-      }
-    }
-  }
-
-  &__shipsWrapper {
-    width: 400px;
-    height: 400px;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    transform: translateY(-40px);
-
-    &-setupShips {
-      width: 100%;
-      display: flex;
-      flex-wrap: wrap;
-      align-items: flex-start;
-    }
-
-    &-ship {
-      background-color: $base-font-color;
-      height: 40px;
-      margin: 5px 10px 5px 0;
-    }
-
-    &__heading {
-      font-weight: 700;
-      width: 100%;
-      display: flex;
-      justify-content: center;
-    }
-  }
-  &__rotateButton {
-    height: 15%;
-    width: 30%;
-    background-color: $btn-color;
-    border: none;
-    border-radius: 15px;
-    font-size: 24px;
-    font-weight: 700;
-    align-self: flex-end;
-    transform: translateY(40px);
-    &:hover {
-      cursor: pointer;
-      background-color: $btn-color-hover;
-    }
-  }
-}
-
-.homaPageButton {
-  border: none;
-  background-color: transparent;
-  color: #fff;
-  font-size: 24px;
-  font-weight: 700;
-  position: absolute;
-  right: 5%;
-  bottom: 3%;
-
-  &:hover {
-    cursor: pointer;
-  }
-}
-
-.startGameButton {
-  height: 10%;
-  width: 20%;
-  background-color: #614029;
-  border: none;
-  border-radius: 15px;
-  font-size: 24px;
-  font-weight: 700;
-  position: absolute;
-  bottom: 3%;
-
-  &:hover {
-    cursor: pointer;
-    background-color: #ad7751;
-  }
-}
