+++ src/game/SHIPS_LIST.js
@@ -1,0 +1,22 @@
+export const SHIPS_LIST = [
+  {
+    shipName: "Carrier",
+    size: 5,
+  },
+  {
+    shipName: "	Battleship",
+    size: 4,
+  },
+  {
+    shipName: "Destroyer",
+    size: 3,
+  },
+  {
+    shipName: "Submarine",
+    size: 2,
+  },
+  {
+    shipName: "Patrol Boat",
+    size: 1,
+  },
+];
