+++ src/variables.scss
@@ -22,3 +22,5 @@ $xl-breakpoint: 1399px;
 $white-color: #ffffff;
 $black-light-color: rgba(0, 0, 0, 0.45);
 $black-dark-color: rgba(0, 0, 0, 0.8);
+$blue-light-color: #243c60;
+$blue-dark-color: #051e42;
